Modify CatalogUpdater to support OAS3 and parse YAML
Previously, OAS3 files wouldn't correctly generate the catalog if the
stage and apiId were embedded in them. Now, they will. Additionally,
YAML used to be passed through as a string into the catalog, but the
frontend (swaggerUI) can't process YAML strings. As such, the catalog
now contains JSON, even if the input is YAML. js-yaml is used to convert
the incoming YAML into JSON swagger or OAS3.