Fix swaggerUI 3 not detecting onChange for Inputs
SwaggerUI 3 uses a library (react-debounce-input) in exactly one place
(the onChange handler for String type inputs in a ParameterRow; where
you type parameters in Try It Out). This library doesn't work with React
16. This commit has the "fixed" assets from a locally built version of
SwaggerUI 3 that doesn't include react-debounce-input.

When/if SwaggerUI 3 merges this in, we can bump the version and not need
to continue floating this change. See this PR:
https://github.com/swagger-api/swagger-ui/pull/4982

WARNING: Building the static assets (say, for a release) WILL overwrite
this bugfix! :/